requirements.txt
# FastAPI core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Pydantic v2 for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Supabase client
supabase>=2.3.0

# HTTP client for async operations (used by supabase)
httpx>=0.25.0

# OpenAI SDK for LLM integration (with OpenRouter)
openai>=1.12.0

# Additional FastAPI dependencies
python-multipart>=0.0.6

# JWT handling
PyJWT>=2.8.0

# Development and testing (optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and monitoring
structlog>=23.2.0 